release 2.0 version

+++ src/backend/agentchat/api/v1/dialog.py
@@ -17,8 +17,9 @@ async def get_dialog(login_user: UserPayload = Depends(get_login_user)):
         results = []
         for msg in messages:
             msg_agent = await AgentService.select_agent_by_id(agent_id=msg["agent_id"])
+
             msg.update(msg_agent)
-            results.extend(msg)
+            results.append(msg)
         return resp_200(data=results)
     except Exception as err:
         logger.error(err)
@@ -26,13 +27,13 @@ async def get_dialog(login_user: UserPayload = Depends(get_login_user)):
 
 
 @router.post("/dialog", response_model=UnifiedResponseModel)
-async def create_dialog(*,
-                        dialog_req: DialogCreateRequest,
+async def create_dialog(dialog_req: DialogCreateRequest = Body(),
                         login_user: UserPayload = Depends(get_login_user)):
     try:
-        await DialogService.create_dialog(name=dialog_req.name, agent_id=dialog_req.agent_id,
-                                          agent_type=dialog_req.agent_type, user_id=login_user.user_id)
-        return resp_200()
+        dialog = await DialogService.create_dialog(name=dialog_req.name, agent_id=dialog_req.agent_id,
+                                                   agent_type=dialog_req.agent_type, user_id=login_user.user_id)
+        print(dialog)
+        return resp_200(dialog)
     except Exception as err:
         logger.error(err)
         return resp_500(message=str(err))
@@ -42,6 +43,9 @@ async def create_dialog(*,
 async def delete_dialog(dialog_id: str = Body(description="对话ID", embed=True),
                         login_user: UserPayload = Depends(get_login_user)):
     try:
+        # 验证用户权限
+        await DialogService.verify_user_permission(dialog_id, login_user.user_id)
+
         await DialogService.delete_dialog(dialog_id=dialog_id)
         return resp_200()
     except Exception as err:
